+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Slit testing framework helpers from acceptance test frameworks. This means we
+  can now use a combination of RSpec/Minitest and Capybara/ActiveSupport test
+  cases.
+
 ## 0.2.0
 
 * Include RSpec + Capybara helper class